Refactored!
