+++ backend/backend.js
@@ -8,14 +8,9 @@ import {
 } from "./kafka";
 import { config, kafka, logs } from "./persistence";
 import { ipcMain } from "electron";
-import { sendUserMessage } from "./messaging";
-import { init as RunTimer } from "./timer";
-import {
-  closeAdmin,
-  initAdmin,
-  getTopicsMeta,
-  deleteTopicFromCluster,
-} from "./kafka/admin/admin";
+import { sendUserMessage } from "./utils/messaging";
+import { init as RunTimer } from "./utils/timer";
+import { closeAdmin, initAdmin, getTopicsMeta  bv} from "./kafka/admin/admin";
 import { kafkaInit } from "./kafka/kafka";
 
 const reconnectKafka = () => {
