+++ dist_electron/index.js
@@ -86,30 +86,6 @@
 /************************************************************************/
 /******/ ({
 
-/***/ "./backend/CustomLogger.js":
-/*!*********************************!*\
-  !*** ./backend/CustomLogger.js ***!
-  \*********************************/
-/*! exports provided: CustomLogCreator */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"CustomLogCreator\", function() { return CustomLogCreator; });\n/* harmony import */ var _persistence_index__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./persistence/index */ \"./backend/persistence/index.js\");\nconst { logLevel } = __webpack_require__(/*! kafkajs */ \"./node_modules/kafkajs/index.js\");\r\n\r\nconst toCustomLogLevel = (level) => {\r\n  switch (level) {\r\n    case logLevel.ERROR:\r\n    case logLevel.NOTHING:\r\n      return \"ERROR\";\r\n    case logLevel.WARN:\r\n      return \"WARN\";\r\n    case logLevel.INFO:\r\n      return \"INFO\";\r\n    case logLevel.DEBUG:\r\n      return \"DEBUG\";\r\n  }\r\n};\r\n\r\nconst CustomLogCreator = () => {\r\n  return ({ level, log }) => {\r\n    const { message } = log;\r\n    Object(_persistence_index__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(message, toCustomLogLevel(level));\r\n  };\r\n\r\n  //   const logger = winston.createLogger({\r\n  //     level: toCustomLogLevel(logLevel),\r\n  //     transports: [\r\n  //         new winston.transports.Console(),\r\n  //         new winston.transports.File({ filename: 'myapp.log' })\r\n  //     ]\r\n  // })\r\n\r\n  // return ({ namespace, level, label, log }) => {\r\n  //     const { message, ...extra } = log\r\n  //     logger.log({\r\n  //         level: toCustomLogLevel(level),\r\n  //         message,\r\n  //         extra,\r\n  //     })\r\n  // }\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/CustomLogger.js?");
-
-/***/ }),
-
-/***/ "./backend/DBSize.js":
-/*!***************************!*\
-  !*** ./backend/DBSize.js ***!
-  \***************************/
-/*! exports provided: dirsizeSync */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"dirsizeSync\", function() { return dirsizeSync; });\nconst path = __webpack_require__(/*! path */ \"path\");\r\nconst fs = __webpack_require__(/*! fs */ \"fs\");\r\n\r\nfunction dirsizeSync(dirname) {\r\n  let size = 0;\r\n  try {\r\n    fs.readdirSync(dirname)\r\n      .map((e) => path.join(dirname, e))\r\n      .map((e) => {\r\n        try {\r\n          return {\r\n            dirname: e,\r\n            stat: fs.statSync(e),\r\n          };\r\n        } catch (ex) {\r\n          return null;\r\n        }\r\n      })\r\n      .forEach((e) => {\r\n        if (e) {\r\n          if (e.stat.isDirectory()) {\r\n            size += dirsizeSync(e.dirname);\r\n          } else if (e.stat.isFile()) {\r\n            size += e.stat.size;\r\n          }\r\n        }\r\n      });\r\n  } catch (ex) {\r\n    console.log(ex);\r\n  }\r\n  return size;\r\n}\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/DBSize.js?");
-
-/***/ }),
-
 /***/ "./backend/backend.js":
 /*!****************************!*\
   !*** ./backend/backend.js ***!
@@ -118,7 +94,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _kafka__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./kafka */ \"./backend/kafka/index.js\");\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! electron */ \"electron\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(electron__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var _messaging__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./messaging */ \"./backend/messaging.js\");\n/* harmony import */ var _timer__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./timer */ \"./backend/timer.js\");\n/* harmony import */ var _kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./kafka/admin/admin */ \"./backend/kafka/admin/admin.js\");\n/* harmony import */ var _kafka_kafka__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./kafka/kafka */ \"./backend/kafka/kafka.js\");\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\nconst reconnectKafka = () => {\r\n  // Stop existing client\r\n  try {\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n    Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n    Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"WARN\", `Disconnnected!`);\r\n  } catch (e) {\r\n    Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to disconnect!`);\r\n  }\r\n  try {\r\n    //Start again\r\n    Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n      .then(() => {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n        Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"initAdmin\"])();\r\n      })\r\n      .catch((err) => {\r\n        console.log(\"Kafka\", err);\r\n      });\r\n    Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Reconnected!`);\r\n  } catch (e) {\r\n    Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to reconnect!`);\r\n  }\r\n};\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"logGet\", (event, arg) => {\r\n  _persistence__WEBPACK_IMPORTED_MODULE_1__[\"logs\"]\r\n    .getlogs(arg)\r\n    .then((logs_) => {\r\n      event.reply(\"logGetResponse\", logs_);\r\n    })\r\n    .catch((err) => {\r\n      console.log(err);\r\n    });\r\n});\r\n\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"kafka\", (event, arg) => {\r\n  var command = arg.command;\r\n\r\n  switch (command) {\r\n    case \"init\":\r\n      Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n        .then(() => {\r\n          Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n          Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n          Object(_timer__WEBPACK_IMPORTED_MODULE_4__[\"init\"])();\r\n          event.reply(\"userMessage\", { type: \"load\", value: true });\r\n        })\r\n        .catch(() => {\r\n          event.reply(\"userMessage\", { type: \"load\", value: false });\r\n        });\r\n      break;\r\n\r\n    case \"close\":\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n      break;\r\n\r\n    case \"message\":\r\n      var payload = arg.payload;\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"SendMessage\"])(payload.topic, payload.key, payload.value, payload.partition)\r\n        .then((res) => {\r\n          if (res) {\r\n            Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Message published sucessfully.\");\r\n          } else {\r\n            Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to publish message.\");\r\n          }\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to publish message.\");\r\n        });\r\n      break;\r\n\r\n    case \"createtopic\":\r\n      var payload_ = arg.payload;\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .addTopic(payload_.name, payload_.type)\r\n        .then((res) => {\r\n          if (res) Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic added to the DB sucessfully\");\r\n          else Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"WARN\", \"Topic already exists!\");\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error adding topic to the DB\");\r\n        });\r\n      if (payload_.createincluseter) {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"createTopic\"])(\r\n          payload_.topic,\r\n          payload_.partition || 1,\r\n          payload_.replicationfactor || 1\r\n        )\r\n          .then(() => {\r\n            Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic created in the cluster.\");\r\n          })\r\n          .catch(() => {\r\n            Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to create topic in the cluster.\");\r\n          });\r\n      }\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"gettopics\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].getTopics().then((topics) => {\r\n        event.reply(\"kafkaResponse\", { topics: topics, type: \"topics\" });\r\n      });\r\n      break;\r\n\r\n    case \"disabletopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .disableTopic(arg.id)\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic disabled sucessfully\");\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error disbaling topic.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"enabletopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .enableTopic(arg.id)\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic enabled sucessfully\");\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error enabling topic.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"removetopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeTopic(arg.id)\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic removed from DB\");\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error while removing topic from the DB.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"getmessages\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .getMessages(arg)\r\n        .then((res) => {\r\n          event.reply(\"kafkamessagesresponse\", res);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed loading kafka messages\");\r\n        });\r\n      break;\r\n\r\n    case \"gettopicsmeta\":\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"getTopicsMeta\"])()\r\n        .then((res) => {\r\n          event.reply(\"kafkaResponse\", { type: \"meta\", meta: res });\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed loading topics from cluster.\");\r\n        });\r\n      return;\r\n    case \"removetopicfromcluster\":\r\n      // deleteTopicFromCluster(arg.topic)\r\n      //   .then(() => {\r\n      //     sendUserMessage(\r\n      //       \"INFO\",\r\n      //       \"Topic removed from the cluster successfully.\"\r\n      //     );\r\n      //   })\r\n      //   .catch(() => {\r\n      //     sendUserMessage(\r\n      //       \"ERROR\",\r\n      //       \"Failed to delete topic. Check if cluster allows it.\"\r\n      //     );\r\n      //   });\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\r\n        \"INFO\",\r\n        \"Feature coming soon.. Topic will be removed from DB if exists!\"\r\n      );\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].removeTopicByName(arg.topic).catch(() => {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to remove topic from the DB.\");\r\n      });\r\n      return;\r\n  }\r\n});\r\n\r\n// Configuration Related.\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"conf\", async (event, arg) => {\r\n  var command = arg.command;\r\n  if (command === \"GET\") {\r\n    event.reply(\"confres\", {\r\n      type: \"GET\",\r\n      server: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_BOOTSTRAP_SERVER\"),\r\n      username: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_USERNAME\"),\r\n      password: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_PASSWORD\"),\r\n      consgrp: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_CONSUMER_GROUP\"),\r\n    });\r\n  }\r\n  if (command === \"SET\") {\r\n    try {\r\n      console.log(arg);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_USERNAME\", arg.username);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_PASSWORD\", arg.password);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_BOOTSTRAP_SERVER\", arg.server);\r\n      reconnectKafka();\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Server config updated sucessfully!`);\r\n    } catch (e) {\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Server config update failed!`);\r\n    }\r\n  }\r\n  if (command === \"STOP\") {\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n    Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n  }\r\n  if (command === \"SET_CONS_GRP\") {\r\n    try {\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_CONSUMER_GROUP\", arg.val);\r\n      reconnectKafka();\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Default consumer group updated sucessfully!`);\r\n    } catch (e) {\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Default consumer group update failed!`);\r\n    }\r\n  }\r\n\r\n  if (command === \"START\") {\r\n    Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n      .then(() => {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n        Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"initAdmin\"])();\r\n        console.log(\"Kafka Initialized\");\r\n      })\r\n      .catch((err) => {\r\n        console.log(\"Kafka\", err);\r\n      });\r\n  }\r\n  if (command === \"PURGE_DATABASE\") {\r\n    if (arg.CONFIG) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"]\r\n        .removeAllConfig()\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `App config cleared!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `App config clear failed!`);\r\n        });\r\n    }\r\n    if (arg.LOGS) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"logs\"]\r\n        .removeAllLogs()\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Logs purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge logs`);\r\n        });\r\n    }\r\n    if (arg.MESSAGES) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeAllMessages()\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Messages purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge messages`);\r\n        });\r\n    }\r\n    if (arg.TOPICS) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeAllTopics()\r\n        .then(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Kafka topics purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge kafka topics.`);\r\n        });\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./backend/backend.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _kafka__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./kafka */ \"./backend/kafka/index.js\");\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! electron */ \"electron\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(electron__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var _utils_messaging__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./utils/messaging */ \"./backend/utils/messaging.js\");\n/* harmony import */ var _utils_timer__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./utils/timer */ \"./backend/utils/timer.js\");\n/* harmony import */ var _kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./kafka/admin/admin */ \"./backend/kafka/admin/admin.js\");\n/* harmony import */ var _kafka_kafka__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./kafka/kafka */ \"./backend/kafka/kafka.js\");\n\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\nconst reconnectKafka = () => {\r\n  // Stop existing client\r\n  try {\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n    Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n    Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"WARN\", `Disconnnected!`);\r\n  } catch (e) {\r\n    Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to disconnect!`);\r\n  }\r\n  try {\r\n    //Start again\r\n    Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n      .then(() => {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n        Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"initAdmin\"])();\r\n      })\r\n      .catch((err) => {\r\n        console.log(\"Kafka\", err);\r\n      });\r\n    Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Reconnected!`);\r\n  } catch (e) {\r\n    Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to reconnect!`);\r\n  }\r\n};\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"logGet\", (event, arg) => {\r\n  _persistence__WEBPACK_IMPORTED_MODULE_1__[\"logs\"]\r\n    .getlogs(arg)\r\n    .then((logs_) => {\r\n      event.reply(\"logGetResponse\", logs_);\r\n    })\r\n    .catch((err) => {\r\n      console.log(err);\r\n    });\r\n});\r\n\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"kafka\", (event, arg) => {\r\n  var command = arg.command;\r\n\r\n  switch (command) {\r\n    case \"init\":\r\n      Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n        .then(() => {\r\n          Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n          Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n          Object(_utils_timer__WEBPACK_IMPORTED_MODULE_4__[\"init\"])();\r\n          event.reply(\"userMessage\", { type: \"load\", value: true });\r\n        })\r\n        .catch(() => {\r\n          event.reply(\"userMessage\", { type: \"load\", value: false });\r\n        });\r\n      break;\r\n\r\n    case \"close\":\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n      break;\r\n\r\n    case \"message\":\r\n      var payload = arg.payload;\r\n      Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"SendMessage\"])(payload.topic, payload.key, payload.value, payload.partition)\r\n        .then((res) => {\r\n          if (res) {\r\n            Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Message published sucessfully.\");\r\n          } else {\r\n            Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to publish message.\");\r\n          }\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to publish message.\");\r\n        });\r\n      break;\r\n\r\n    case \"createtopic\":\r\n      var payload_ = arg.payload;\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .addTopic(payload_.name, payload_.type)\r\n        .then((res) => {\r\n          if (res) Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic added to the DB sucessfully\");\r\n          else Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"WARN\", \"Topic already exists!\");\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error adding topic to the DB\");\r\n        });\r\n      if (payload_.createincluseter) {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"createTopic\"])(\r\n          payload_.topic,\r\n          payload_.partition || 1,\r\n          payload_.replicationfactor || 1\r\n        )\r\n          .then(() => {\r\n            Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic created in the cluster.\");\r\n          })\r\n          .catch(() => {\r\n            Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to create topic in the cluster.\");\r\n          });\r\n      }\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"gettopics\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].getTopics().then((topics) => {\r\n        event.reply(\"kafkaResponse\", { topics: topics, type: \"topics\" });\r\n      });\r\n      break;\r\n\r\n    case \"disabletopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .disableTopic(arg.id)\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic disabled sucessfully\");\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error disbaling topic.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"enabletopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .enableTopic(arg.id)\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic enabled sucessfully\");\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error enabling topic.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"removetopic\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeTopic(arg.id)\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", \"Topic removed from DB\");\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Error while removing topic from the DB.\");\r\n        });\r\n      reconnectKafka();\r\n      break;\r\n\r\n    case \"getmessages\":\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .getMessages(arg)\r\n        .then((res) => {\r\n          event.reply(\"kafkamessagesresponse\", res);\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed loading kafka messages\");\r\n        });\r\n      break;\r\n\r\n    case \"gettopicsmeta\":\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"getTopicsMeta\"])()\r\n        .then((res) => {\r\n          event.reply(\"kafkaResponse\", { type: \"meta\", meta: res });\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed loading topics from cluster.\");\r\n        });\r\n      return;\r\n    case \"removetopicfromcluster\":\r\n      // deleteTopicFromCluster(arg.topic)\r\n      //   .then(() => {\r\n      //     sendUserMessage(\r\n      //       \"INFO\",\r\n      //       \"Topic removed from the cluster successfully.\"\r\n      //     );\r\n      //   })\r\n      //   .catch(() => {\r\n      //     sendUserMessage(\r\n      //       \"ERROR\",\r\n      //       \"Failed to delete topic. Check if cluster allows it.\"\r\n      //     );\r\n      //   });\r\n      Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\r\n        \"INFO\",\r\n        \"Feature coming soon.. Topic will be removed from DB if exists!\"\r\n      );\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].removeTopicByName(arg.topic).catch(() => {\r\n        Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", \"Failed to remove topic from the DB.\");\r\n      });\r\n      return;\r\n  }\r\n});\r\n\r\n// Configuration Related.\r\nelectron__WEBPACK_IMPORTED_MODULE_2__[\"ipcMain\"].on(\"conf\", async (event, arg) => {\r\n  var command = arg.command;\r\n  if (command === \"GET\") {\r\n    event.reply(\"confres\", {\r\n      type: \"GET\",\r\n      server: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_BOOTSTRAP_SERVER\"),\r\n      username: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_USERNAME\"),\r\n      password: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_PASSWORD\"),\r\n      consgrp: await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].readConfig(\"KAFKA_CONSUMER_GROUP\"),\r\n    });\r\n  }\r\n  if (command === \"SET\") {\r\n    try {\r\n      console.log(arg);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_USERNAME\", arg.username);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_PASSWORD\", arg.password);\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_BOOTSTRAP_SERVER\", arg.server);\r\n      reconnectKafka();\r\n      Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Server config updated sucessfully!`);\r\n    } catch (e) {\r\n      Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Server config update failed!`);\r\n    }\r\n  }\r\n  if (command === \"STOP\") {\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerClose\"])();\r\n    Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerClose\"])();\r\n    Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"closeAdmin\"])();\r\n  }\r\n  if (command === \"SET_CONS_GRP\") {\r\n    try {\r\n      await _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"].updateConfig(\"KAFKA_CONSUMER_GROUP\", arg.val);\r\n      reconnectKafka();\r\n      Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Default consumer group updated sucessfully!`);\r\n    } catch (e) {\r\n      Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Default consumer group update failed!`);\r\n    }\r\n  }\r\n\r\n  if (command === \"START\") {\r\n    Object(_kafka_kafka__WEBPACK_IMPORTED_MODULE_6__[\"kafkaInit\"])()\r\n      .then(() => {\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ConsumerInit\"])();\r\n        Object(_kafka__WEBPACK_IMPORTED_MODULE_0__[\"ProducerInit\"])();\r\n        Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_5__[\"initAdmin\"])();\r\n        console.log(\"Kafka Initialized\");\r\n      })\r\n      .catch((err) => {\r\n        console.log(\"Kafka\", err);\r\n      });\r\n  }\r\n  if (command === \"PURGE_DATABASE\") {\r\n    if (arg.CONFIG) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"config\"]\r\n        .removeAllConfig()\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `App config cleared!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `App config clear failed!`);\r\n        });\r\n    }\r\n    if (arg.LOGS) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"logs\"]\r\n        .removeAllLogs()\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Logs purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge logs`);\r\n        });\r\n    }\r\n    if (arg.MESSAGES) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeAllMessages()\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Messages purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge messages`);\r\n        });\r\n    }\r\n    if (arg.TOPICS) {\r\n      _persistence__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .removeAllTopics()\r\n        .then(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"INFO\", `Kafka topics purged successfully!`);\r\n        })\r\n        .catch(() => {\r\n          Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendUserMessage\"])(\"ERROR\", `Failed to purge kafka topics.`);\r\n        });\r\n    }\r\n  }\r\n});\r\n\n\n//# sourceURL=webpack:///./backend/backend.js?");
 
 /***/ }),
 
@@ -142,7 +118,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"init\", function() { return init; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"close\", function() { return close; });\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var _messaging__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../messaging */ \"./backend/messaging.js\");\n\r\n\r\n\r\nconst init = async () => {\r\n  // Create a Kafka consumer.\r\n  var consumerGroup = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_CONSUMER_GROUP\");\r\n  console.log(\"consumerGroup\", consumerGroup);\r\n  global.consumer = global.kafka.consumer({\r\n    groupId:\r\n      consumerGroup && consumerGroup !== \"\" ? consumerGroup : \"KAFKA_EXPLORER\",\r\n  });\r\n\r\n  const topics = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"kafka\"].getKafkaTopicsToConsume();\r\n\r\n  // Connect\r\n  await global.consumer.connect();\r\n\r\n  // Subscribe to Kafka topic\r\n  topics.map(async (topic) => {\r\n    if (topic.isActive) {\r\n      // console.log(`Subscribing to topic ${topic.name}`, \"INFO\");\r\n      Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(`Subscribing to topic ${topic.name}`, \"INFO\");\r\n      try {\r\n        await global.consumer.subscribe({\r\n          topic: topic.name,\r\n          fromBeginning: false,\r\n        });\r\n      } catch (e) {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_1__[\"sendUserMessage\"])(\r\n          \"ERROR\",\r\n          `Failed to subscribe to topic : ${topic.name}`\r\n        );\r\n      }\r\n    }\r\n  });\r\n\r\n  // Listen to incoming messages\r\n  await global.consumer.run({\r\n    eachMessage: ({ topic, partition, message }) => {\r\n      var msg = {\r\n        key: message.key ? message.key.toString() : \"\",\r\n        value: message.value ? message.value.toString() : \"\",\r\n        headers: message.headers ? message.headers : \"\",\r\n      };\r\n      // console.log(JSON.stringify(msg));\r\n      _persistence__WEBPACK_IMPORTED_MODULE_0__[\"kafka\"]\r\n        .addMessage(JSON.stringify(msg), topic, \"consume\", partition, true)\r\n        .catch(async (error) => {\r\n          Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(\r\n            `Failed to store incoming message ${topic} Error: ${error}`,\r\n            \"WARN\"\r\n          );\r\n        });\r\n    },\r\n  });\r\n};\r\n\r\nconst close = async () => {\r\n  try {\r\n    await global.consumer.disconnect();\r\n  } catch (e) {\r\n    Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(`Failed to disconnect consumer gracefully!`, \"ERROR\");\r\n  }\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/kafka/consumer/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"init\", function() { return init; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"close\", function() { return close; });\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var _utils_messaging__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/messaging */ \"./backend/utils/messaging.js\");\n\r\n\r\n\r\nconst init = async () => {\r\n  // Create a Kafka consumer.\r\n  var consumerGroup = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_CONSUMER_GROUP\");\r\n  console.log(\"consumerGroup\", consumerGroup);\r\n  global.consumer = global.kafka.consumer({\r\n    groupId:\r\n      consumerGroup && consumerGroup !== \"\" ? consumerGroup : \"KAFKA_EXPLORER\",\r\n  });\r\n\r\n  const topics = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"kafka\"].getKafkaTopicsToConsume();\r\n\r\n  // Connect\r\n  await global.consumer.connect();\r\n\r\n  // Subscribe to Kafka topic\r\n  topics.map(async (topic) => {\r\n    if (topic.isActive) {\r\n      // console.log(`Subscribing to topic ${topic.name}`, \"INFO\");\r\n      Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(`Subscribing to topic ${topic.name}`, \"INFO\");\r\n      try {\r\n        await global.consumer.subscribe({\r\n          topic: topic.name,\r\n          fromBeginning: false,\r\n        });\r\n      } catch (e) {\r\n        Object(_utils_messaging__WEBPACK_IMPORTED_MODULE_1__[\"sendUserMessage\"])(\r\n          \"ERROR\",\r\n          `Failed to subscribe to topic : ${topic.name}`\r\n        );\r\n      }\r\n    }\r\n  });\r\n\r\n  // Listen to incoming messages\r\n  await global.consumer.run({\r\n    eachMessage: ({ topic, partition, message }) => {\r\n      var msg = {\r\n        key: message.key ? message.key.toString() : \"\",\r\n        value: message.value ? message.value.toString() : \"\",\r\n        headers: message.headers ? message.headers : \"\",\r\n      };\r\n      // console.log(JSON.stringify(msg));\r\n      _persistence__WEBPACK_IMPORTED_MODULE_0__[\"kafka\"]\r\n        .addMessage(JSON.stringify(msg), topic, \"consume\", partition, true)\r\n        .catch(async (error) => {\r\n          Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(\r\n            `Failed to store incoming message ${topic} Error: ${error}`,\r\n            \"WARN\"\r\n          );\r\n        });\r\n    },\r\n  });\r\n};\r\n\r\nconst close = async () => {\r\n  try {\r\n    await global.consumer.disconnect();\r\n  } catch (e) {\r\n    Object(_persistence__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(`Failed to disconnect consumer gracefully!`, \"ERROR\");\r\n  }\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/kafka/consumer/index.js?");
 
 /***/ }),
 
@@ -166,7 +142,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _con
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"kafkaInit\", function() { return kafkaInit; });\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var _CustomLogger__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../CustomLogger */ \"./backend/CustomLogger.js\");\nconst { Kafka } = __webpack_require__(/*! kafkajs */ \"./node_modules/kafkajs/index.js\");\r\n\r\n\r\n\r\nconst kafkaInit = () => {\r\n  global.kafka = null;\r\n  let username, password, brokers, sasl, ssl;\r\n  return new Promise((resolve, reject) => {\r\n    setTimeout(() => {\r\n      if (global.kafka === null) reject(new Error(\"Failed to connect!\"));\r\n      else resolve(\"Connected!\");\r\n    }, 15000);\r\n    global.kafka = new Kafka({\r\n      clientId: \"a-random-client-id\",\r\n      brokers: async () => {\r\n        username = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_USERNAME\");\r\n        password = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_PASSWORD\");\r\n        brokers = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_BOOTSTRAP_SERVER\");\r\n        sasl =\r\n          username && password\r\n            ? { username, password, mechanism: \"plain\" }\r\n            : null;\r\n        ssl = !!sasl;\r\n        brokers = brokers.split(\",\");\r\n        return brokers;\r\n      },\r\n      ssl,\r\n      sasl,\r\n      requestTimeout: 25000,\r\n      logCreator: _CustomLogger__WEBPACK_IMPORTED_MODULE_1__[\"CustomLogCreator\"],\r\n    });\r\n  });\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/kafka/kafka.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"kafkaInit\", function() { return kafkaInit; });\n/* harmony import */ var _persistence__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../persistence */ \"./backend/persistence/index.js\");\n/* harmony import */ var _utils_CustomLogger__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../utils/CustomLogger */ \"./backend/utils/CustomLogger.js\");\nconst { Kafka } = __webpack_require__(/*! kafkajs */ \"./node_modules/kafkajs/index.js\");\r\n\r\n\r\n\r\nconst kafkaInit = () => {\r\n  global.kafka = null;\r\n  let username, password, brokers, sasl, ssl;\r\n  return new Promise((resolve, reject) => {\r\n    setTimeout(() => {\r\n      if (global.kafka === null) reject(new Error(\"Failed to connect!\"));\r\n      else resolve(\"Connected!\");\r\n    }, 15000);\r\n    global.kafka = new Kafka({\r\n      clientId: \"a-random-client-id\",\r\n      brokers: async () => {\r\n        username = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_USERNAME\");\r\n        password = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_PASSWORD\");\r\n        brokers = await _persistence__WEBPACK_IMPORTED_MODULE_0__[\"config\"].readConfig(\"KAFKA_BOOTSTRAP_SERVER\");\r\n        sasl =\r\n          username && password\r\n            ? { username, password, mechanism: \"plain\" }\r\n            : null;\r\n        ssl = !!sasl;\r\n        brokers = brokers.split(\",\");\r\n        return brokers;\r\n      },\r\n      ssl,\r\n      sasl,\r\n      requestTimeout: 25000,\r\n      logCreator: _utils_CustomLogger__WEBPACK_IMPORTED_MODULE_1__[\"CustomLogCreator\"],\r\n    });\r\n  });\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/kafka/kafka.js?");
 
 /***/ }),
 
@@ -194,18 +170,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 
 /***/ }),
 
-/***/ "./backend/messaging.js":
-/*!******************************!*\
-  !*** ./backend/messaging.js ***!
-  \******************************/
-/*! exports provided: sendUserMessage, sendCluseterDetails, sendKafkaMessages, sendKafkaTopics, sendLogs, sendDbSize */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendUserMessage\", function() { return sendUserMessage; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendCluseterDetails\", function() { return sendCluseterDetails; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendKafkaMessages\", function() { return sendKafkaMessages; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendKafkaTopics\", function() { return sendKafkaTopics; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendLogs\", function() { return sendLogs; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendDbSize\", function() { return sendDbSize; });\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! electron */ \"electron\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(electron__WEBPACK_IMPORTED_MODULE_0__);\n\r\n\r\nconst sendUserMessage = (type, message) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"userMessage\", {\r\n      type: type,\r\n      message: message,\r\n    });\r\n  });\r\n};\r\n\r\nconst sendCluseterDetails = (data) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"clusterdata\", data);\r\n  });\r\n};\r\n\r\nconst sendKafkaMessages = (data) => {\r\n  let data_ = {};\r\n  data_.type = \"auto\";\r\n  data_.data = data;\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"kafkamessagesresponse\", data_);\r\n  });\r\n};\r\n\r\nconst sendKafkaTopics = (data) => {\r\n  let data_ = {};\r\n  data_.type = \"auto\";\r\n  data_.data = data;\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"kafkatopics\", data_);\r\n  });\r\n};\r\nconst sendLogs = (data) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"autologs\", data);\r\n  });\r\n};\r\n\r\nconst sendDbSize = (size) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"dbsize\", size);\r\n  });\r\n};\r\n\r\n\n\n//# sourceURL=webpack:///./backend/messaging.js?");
-
-/***/ }),
-
 /***/ "./backend/persistence/config.js":
 /*!***************************************!*\
   !*** ./backend/persistence/config.js ***!
@@ -230,15 +194,51 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 
 /***/ }),
 
-/***/ "./backend/timer.js":
-/*!**************************!*\
-  !*** ./backend/timer.js ***!
-  \**************************/
+/***/ "./backend/utils/CustomLogger.js":
+/*!***************************************!*\
+  !*** ./backend/utils/CustomLogger.js ***!
+  \***************************************/
+/*! exports provided: CustomLogCreator */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"CustomLogCreator\", function() { return CustomLogCreator; });\n/* harmony import */ var _persistence_index__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../persistence/index */ \"./backend/persistence/index.js\");\nconst { logLevel } = __webpack_require__(/*! kafkajs */ \"./node_modules/kafkajs/index.js\");\r\n\r\nconst toCustomLogLevel = (level) => {\r\n  switch (level) {\r\n    case logLevel.ERROR:\r\n    case logLevel.NOTHING:\r\n      return \"ERROR\";\r\n    case logLevel.WARN:\r\n      return \"WARN\";\r\n    case logLevel.INFO:\r\n      return \"INFO\";\r\n    case logLevel.DEBUG:\r\n      return \"DEBUG\";\r\n  }\r\n};\r\n\r\nconst CustomLogCreator = () => {\r\n  return ({ level, log }) => {\r\n    const { message } = log;\r\n    Object(_persistence_index__WEBPACK_IMPORTED_MODULE_0__[\"log\"])(message, toCustomLogLevel(level));\r\n  };\r\n\r\n  //   const logger = winston.createLogger({\r\n  //     level: toCustomLogLevel(logLevel),\r\n  //     transports: [\r\n  //         new winston.transports.Console(),\r\n  //         new winston.transports.File({ filename: 'myapp.log' })\r\n  //     ]\r\n  // })\r\n\r\n  // return ({ namespace, level, label, log }) => {\r\n  //     const { message, ...extra } = log\r\n  //     logger.log({\r\n  //         level: toCustomLogLevel(level),\r\n  //         message,\r\n  //         extra,\r\n  //     })\r\n  // }\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/utils/CustomLogger.js?");
+
+/***/ }),
+
+/***/ "./backend/utils/DBSize.js":
+/*!*********************************!*\
+  !*** ./backend/utils/DBSize.js ***!
+  \*********************************/
+/*! exports provided: dirsizeSync */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"dirsizeSync\", function() { return dirsizeSync; });\nconst path = __webpack_require__(/*! path */ \"path\");\r\nconst fs = __webpack_require__(/*! fs */ \"fs\");\r\n\r\nfunction dirsizeSync(dirname) {\r\n  let size = 0;\r\n  try {\r\n    fs.readdirSync(dirname)\r\n      .map((e) => path.join(dirname, e))\r\n      .map((e) => {\r\n        try {\r\n          return {\r\n            dirname: e,\r\n            stat: fs.statSync(e),\r\n          };\r\n        } catch (ex) {\r\n          return null;\r\n        }\r\n      })\r\n      .forEach((e) => {\r\n        if (e) {\r\n          if (e.stat.isDirectory()) {\r\n            size += dirsizeSync(e.dirname);\r\n          } else if (e.stat.isFile()) {\r\n            size += e.stat.size;\r\n          }\r\n        }\r\n      });\r\n  } catch (ex) {\r\n    console.log(ex);\r\n  }\r\n  return size;\r\n}\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/utils/DBSize.js?");
+
+/***/ }),
+
+/***/ "./backend/utils/messaging.js":
+/*!************************************!*\
+  !*** ./backend/utils/messaging.js ***!
+  \************************************/
+/*! exports provided: sendUserMessage, sendCluseterDetails, sendKafkaMessages, sendKafkaTopics, sendLogs, sendDbSize */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendUserMessage\", function() { return sendUserMessage; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendCluseterDetails\", function() { return sendCluseterDetails; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendKafkaMessages\", function() { return sendKafkaMessages; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendKafkaTopics\", function() { return sendKafkaTopics; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendLogs\", function() { return sendLogs; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"sendDbSize\", function() { return sendDbSize; });\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! electron */ \"electron\");\n/* harmony import */ var electron__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(electron__WEBPACK_IMPORTED_MODULE_0__);\n\r\n\r\nconst sendUserMessage = (type, message) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"userMessage\", {\r\n      type: type,\r\n      message: message,\r\n    });\r\n  });\r\n};\r\n\r\nconst sendCluseterDetails = (data) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"clusterdata\", data);\r\n  });\r\n};\r\n\r\nconst sendKafkaMessages = (data) => {\r\n  let data_ = {};\r\n  data_.type = \"auto\";\r\n  data_.data = data;\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"kafkamessagesresponse\", data_);\r\n  });\r\n};\r\n\r\nconst sendKafkaTopics = (data) => {\r\n  let data_ = {};\r\n  data_.type = \"auto\";\r\n  data_.data = data;\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"kafkatopics\", data_);\r\n  });\r\n};\r\nconst sendLogs = (data) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"autologs\", data);\r\n  });\r\n};\r\n\r\nconst sendDbSize = (size) => {\r\n  electron__WEBPACK_IMPORTED_MODULE_0__[\"webContents\"].getAllWebContents().map((content) => {\r\n    content.send(\"dbsize\", size);\r\n  });\r\n};\r\n\r\n\n\n//# sourceURL=webpack:///./backend/utils/messaging.js?");
+
+/***/ }),
+
+/***/ "./backend/utils/timer.js":
+/*!********************************!*\
+  !*** ./backend/utils/timer.js ***!
+  \********************************/
 /*! exports provided: init */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"init\", function() { return init; });\n/* harmony import */ var _kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./kafka/admin/admin */ \"./backend/kafka/admin/admin.js\");\n/* harmony import */ var _persistence_index__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./persistence/index */ \"./backend/persistence/index.js\");\n/* harmony import */ var _DBSize__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./DBSize */ \"./backend/DBSize.js\");\n/* harmony import */ var _messaging__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./messaging */ \"./backend/messaging.js\");\n\r\n\r\n\r\n\r\n\r\nconst init = () => {\r\n  Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__[\"initAdmin\"])();\r\n  setInterval(() => {\r\n    try {\r\n      // Send latest messages\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .getMessages({\r\n          start: new Date(new Date() - 24 * 1000 * 60 * 60),\r\n          end: new Date(),\r\n        })\r\n        .then((res) => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendKafkaMessages\"])(res);\r\n        });\r\n\r\n      // Senda all kafka topics\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].getTopics().then((res) => {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendKafkaTopics\"])(res);\r\n      });\r\n\r\n      // send logs latest\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"logs\"].getLatestLogs(6).then((res) => {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendLogs\"])(res);\r\n      });\r\n    } catch (e) {\r\n      Object(_persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"log\"])(`Error occured in timer!` + e, \"ERROR\");\r\n    }\r\n  }, 5000);\r\n  setInterval(() => {\r\n    try {\r\n      var dbSize = Object(_DBSize__WEBPACK_IMPORTED_MODULE_2__[\"dirsizeSync\"])(\"kafka_explorer\");\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendDbSize\"])(dbSize);\r\n      \r\n      // Send Cluster Info\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__[\"getClusterInfo\"])()\r\n        .then((res) => {\r\n          res.timestamp = new Date();\r\n          res.status = true;\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendCluseterDetails\"])(res);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendCluseterDetails\"])({\r\n            timestamp: new Date(),\r\n            status: false,\r\n            brokers: [\r\n              {\r\n                host: \"Not connected\",\r\n                nodeId: -1,\r\n              },\r\n            ],\r\n            controller: -1,\r\n          });\r\n        });\r\n    } catch (e) {\r\n      Object(_persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"log\"])(`Error occured in timer!` + e, \"ERROR\");\r\n    }\r\n  }, 30000);\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/timer.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"init\", function() { return init; });\n/* harmony import */ var _kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../kafka/admin/admin */ \"./backend/kafka/admin/admin.js\");\n/* harmony import */ var _persistence_index__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../persistence/index */ \"./backend/persistence/index.js\");\n/* harmony import */ var _DBSize__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./DBSize */ \"./backend/utils/DBSize.js\");\n/* harmony import */ var _messaging__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./messaging */ \"./backend/utils/messaging.js\");\n\r\n\r\n\r\n\r\n\r\nconst init = () => {\r\n  Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__[\"initAdmin\"])();\r\n  setInterval(() => {\r\n    try {\r\n      // Send latest messages\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"]\r\n        .getMessages({\r\n          start: new Date(new Date() - 24 * 1000 * 60 * 60),\r\n          end: new Date(),\r\n        })\r\n        .then((res) => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendKafkaMessages\"])(res);\r\n        });\r\n\r\n      // Senda all kafka topics\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"kafka\"].getTopics().then((res) => {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendKafkaTopics\"])(res);\r\n      });\r\n\r\n      // send logs latest\r\n      _persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"logs\"].getLatestLogs(6).then((res) => {\r\n        Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendLogs\"])(res);\r\n      });\r\n    } catch (e) {\r\n      Object(_persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"log\"])(`Error occured in timer!` + e, \"ERROR\");\r\n    }\r\n  }, 5000);\r\n  setInterval(() => {\r\n    try {\r\n      var dbSize = Object(_DBSize__WEBPACK_IMPORTED_MODULE_2__[\"dirsizeSync\"])(\"kafka_explorer\");\r\n      Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendDbSize\"])(dbSize);\r\n      \r\n      // Send Cluster Info\r\n      Object(_kafka_admin_admin__WEBPACK_IMPORTED_MODULE_0__[\"getClusterInfo\"])()\r\n        .then((res) => {\r\n          res.timestamp = new Date();\r\n          res.status = true;\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendCluseterDetails\"])(res);\r\n        })\r\n        .catch(() => {\r\n          Object(_messaging__WEBPACK_IMPORTED_MODULE_3__[\"sendCluseterDetails\"])({\r\n            timestamp: new Date(),\r\n            status: false,\r\n            brokers: [\r\n              {\r\n                host: \"Not connected\",\r\n                nodeId: -1,\r\n              },\r\n            ],\r\n            controller: -1,\r\n          });\r\n        });\r\n    } catch (e) {\r\n      Object(_persistence_index__WEBPACK_IMPORTED_MODULE_1__[\"log\"])(`Error occured in timer!` + e, \"ERROR\");\r\n    }\r\n  }, 30000);\r\n};\r\n\r\n\r\n\n\n//# sourceURL=webpack:///./backend/utils/timer.js?");
 
 /***/ }),
